feat(Game): implement lock-free data pipeline for item definitions
- Added CConfigReader for overlapping disk I/O operations via ICryPak.
- Implemented CItemParser with RapidJSON In-Situ parsing for zero-copy deserialization.
- Created CItemLoader using CryEngine IJobManager to parallelize JSON parsing.
- Added CItemDatabase with batch registration to prevent unordered_map rehashing.
- Implemented Prototype pattern (Clone) for IItemTrait.

+++ Game/Code/Systems/Items/Item.h
@@ -3,7 +3,7 @@
 #include <memory>
 #include <vector>
 
-using TItemID = uint32_t;
+using TItemID = uint64_t;
 
 class CItem
 {
